Better like this

+++ code/modules/mob/new_player/new_player.dm
@@ -384,8 +384,9 @@
 	ticker.mode.latespawn(character)//can we make them a latejoin antag?
 
 	if (!character || !character.mind) //Character got transformed in a latejoin ruleset
+		if(character)
+			qdel(character)
 		qdel(src)
-		qdel(character)
 		return
 
 	// Very hacky. Sorry about that
